Formats SendMessage component
Formats the SendMessage component to resemble Mihoa's design.

Uses a textarea instead of an input and autosizes to fit the size of the message being sent. Also adds a max-length of 300 characters per message.

Implements a very simple fix to keep components from running of the page by setting the max-width of the root element to 90vw.