ChatBox Component fixes and features.
Fixes #6

Adds display flex to main body of page so that the ChatBox component takes up the majority of the screen with the SendMessage at the bottom and NavBar at the top.

We will probably need to add Flex: 1 to the css of the Welcome component as well so it works with this layout.

Also makes it so scrolling to the top all the messages will load 50 more, until you've reach 1000 messages or all the messages in the database (whichever is less).

Still would like to add some sort of notification when new messages show up that you haven't read. Maybe also a message for when you've loaded all the messages and there are no more to load.